moved pathbuilder class
